refactor: extract set command mixin from coordinator core
